Move KeyStorageStatus to assoc data classes
